-deprecate old QR scanner - use new QR scanner + rename
